+++ packages/@headlessui-react/src/components/dialog/dialog.tsx
@@ -299,7 +299,14 @@ function DialogFn<TTag extends ElementType = typeof DEFAULT_DIALOG_TAG>(
     if (hasNestedDialogs) return false
     return true
   })()
-  useOutsideClick(resolveRootContainers, close, outsideClickEnabled)
+  useOutsideClick(
+    resolveRootContainers,
+    (event) => {
+      event.preventDefault()
+      close()
+    },
+    outsideClickEnabled
+  )
 
   // Handle `Escape` to close
   let escapeToCloseEnabled = (() => {